FM-34: Split the fvm interpreter into smaller modules.
